+++ parlai/scripts/convert_data_to_parlai_format.py
@@ -19,7 +19,7 @@ from parlai.agents.repeat_label.repeat_label import RepeatLabelAgent
 from parlai.core.worlds import create_task
 from parlai.utils.misc import msg_to_str, TimeLogger
 import parlai.utils.logging as logging
-from parlai.scripts.script import ParlaiScript
+from parlai.core.script import ParlaiScript, register_script
 import random
 import tempfile
