+++ parlai/scripts/distributed_train.py
@@ -38,7 +38,7 @@ import subprocess
 import parlai.scripts.train_model as single_train
 import parlai.utils.logging as logging
 from parlai.scripts.multiprocessing_train import multiprocess_train
-from parlai.scripts.script import ParlaiScript
+from parlai.core.script import ParlaiScript, register_script
 
 
 def setup_args():
